[FEAT] - Lib + policy + controller + redirect fix

+++ app/Http/Controllers/Assets.php
@@ -31,12 +31,21 @@ use App\Models\Attachment;
  */
 class Assets extends Controller
 {
+
+    public function __construct() {
+        $this->authorizeResource(Asset::class, 'asset');
+    }
+
     /**
      * Display a listing of the assets.
      * 
      * @return Response
      */
     public function index(): Response {
+
+        // Authorization
+        $this->authorize('index', Asset::class);
+
         return Inertia::render('assets/index', ['assets' => Asset::getTreeOrderedAssets()]);
     }
 
@@ -46,6 +55,9 @@ class Assets extends Controller
      * @return Response
      */
     public function create(): Response {
+
+        // Authorization
+        $this->authorize('create', Asset::class);
 
         $assets = Asset::getTreeOrderedAssets();
 
@@ -66,6 +78,10 @@ class Assets extends Controller
      * @return Response | RedirectResponse
      */
     public function edit(Asset $asset): Response | RedirectResponse {
+
+        // Authorization
+        $this->authorize('edit', $asset);
+
         if(!$asset) return redirect()->route('assets.index')->with(['error' => __('Asset doesn\'t exist')]);
 
         $assets = Asset::getTreeOrderedAssets();
@@ -87,6 +103,10 @@ class Assets extends Controller
      * @return Response | RedirectResponse
      */
     public function show(Asset $asset): Response | RedirectResponse {
+
+        // Authorization
+        $this->authorize('show', $asset);
+
         if(!$asset) return redirect()->route('assets.index')->with(['error' => __('Asset doesn\'t exist')]);
         return Inertia::render('assets/show', ['asset' => $asset->load('attachments'), 'assets' => Asset::getTreeOrderedAssets()]);
     }
@@ -98,6 +118,10 @@ class Assets extends Controller
      * @return RedirectResponse
      */
     public function store(RequestsStore $request): RedirectResponse {
+
+        // Authorization
+        $this->authorize('store', Asset::class);
+
         $data = $request->validated();
 
         $asset = Asset::create([
@@ -157,6 +181,9 @@ class Assets extends Controller
      * @return RedirectResponse
      */
     public function update(RequestsUpdate $request, Asset $asset): RedirectResponse {
+
+        // Authorization
+        $this->authorize('update', $asset);
         
         $data = $request->validated();
 
@@ -261,6 +288,10 @@ class Assets extends Controller
      * @return RedirectResponse
      */
     public function destroy(Asset $asset): RedirectResponse {
+
+        // Authorization
+        $this->authorize('delete', $asset);
+
         $asset->delete();
         return redirect()->route('assets.index')->with(['success' => __('Asset deleted successfully.')]);
     }
@@ -272,6 +303,10 @@ class Assets extends Controller
      * @return RedirectResponse
      */
     public function restore(Asset $asset): RedirectResponse {
+        
+        // Authorization
+        $this->authorize('restore', $asset);
+        
         $asset->restore();
         return redirect()->route('assets.index')->with(['success' => __('Asset restored successfully.')]);
     }
@@ -283,6 +318,10 @@ class Assets extends Controller
      * @return RedirectResponse
      */
     public function forceDelete(Asset $asset): RedirectResponse {
+
+        // Authorization
+        $this->authorize('forceDelete', $asset);
+
         $asset->forceDelete();
         return redirect()->route('assets.index')->with(['success' => __('Asset permanently deleted successfully.')]);
     }
